% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{ac_simu_wrapper}
\alias{ac_simu_wrapper}
\title{Run the age-cohort model with ridge regularization on simulated data}
\usage{
ac_simu_wrapper(ind_wrapper, design, exhaust_ls, scaled_ground_haz)
}
\description{
Run the age-cohort model with ridge regularization on simulated data
}
\seealso{
Other simu_wrapper: \code{\link{criterion_simu_wrapper}},
  \code{\link{cv_simu_wrapper}}, \code{\link{design_simu}},
  \code{\link{ridge_simu_wrapper}}
}
\concept{simu_wrapper}
