% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{band_prod}
\alias{band_prod}
\title{Solve linear system when the matrix is block-wise with a band diagonal block}
\usage{
band_prod(mat, vect, mat_as_rotated = FALSE, lim = NULL)
}
\arguments{
\item{mat}{Square matrix}

\item{vect}{Vector of same length as \code{mat}}

\item{lim}{Positive integer}
}
\value{
The vector \code{mat} ^ {-1} \code{vect} is returned.
If \code{mat[lim:end, lim:end]} is not a band matrix or if \code{mat} is
symmetric, an error is returned.
If these two conditions are met, the returned value is the same as
\code{solve(mat, vect)}. There might be nummerical impercision errors.
}
\description{
Solve linear system when the matrix is block-wise with a band diagonal block
}
\examples{
## With mat_as_rotated = FALSE
set.seed(0)
A <- matrix(rnorm(36), 6, 6)
B <- matrix(rnorm(24), 6, 4)
D <- diag(rnorm(4)) + rbind(cbind(0, diag(rnorm(3))), 0)
mat <- rbind(cbind(A + t(A), B),
                     cbind(t(B), D + t(D)))
vect <- rnorm(10)
lim <- 6
max(band_prod(mat, vect, lim) - solve(mat, vect))

## With mat_as_rotated = TRUE
set.seed(0)
A <- matrix(rnorm(36), 6, 6)
B <- matrix(rnorm(24), 6, 4)
D <- cbind(rnorm(4), c(rnorm(3), 0))
mat_rotated <- list('A' = A + t(A),
            'B' = B,
            'D' = D)
mat <- rbind(cbind(mat_rotated$A, mat_rotated$B),
                     cbind(t(mat_rotated$B), rot2mat(mat_rotated$D)))
vect <- rnorm(10)
max(band_prod(mat_rotated, vect, mat_as_rotated = TRUE) - solve(mat, vect))
}
