% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chew}
\alias{chew}
\title{Power Exponentiated Weibull (EW) cumulative hazard function.
https://rpubs.com/FJRubio/EWD}
\usage{
chew(t, sigma, nu, gamma)
}
\arguments{
\item{t}{: positive argument}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}
}
\value{
the value of the EW cumulative hazard function
}
\description{
Power Exponentiated Weibull (EW) cumulative hazard function.
https://rpubs.com/FJRubio/EWD
}
