% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chgamma}
\alias{chgamma}
\title{Gamma (G) cumulative hazard function.}
\usage{
chgamma(t, shape, scale)
}
\arguments{
\item{t}{: positive argument}

\item{shape}{: shape parameter}

\item{scale}{: scale parameter}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the Weibull hazard function
}
\description{
Gamma (G) cumulative hazard function.
}
