% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chggamma}
\alias{chggamma}
\title{Generalised Gamma (GG) cumulative hazard function.
https://rpubs.com/FJRubio/GG}
\usage{
chggamma(t, sigma, nu, gamma)
}
\arguments{
\item{t}{: positive argument}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}
}
\value{
the value of the GG cumulative hazard function
}
\description{
Generalised Gamma (GG) cumulative hazard function.
https://rpubs.com/FJRubio/GG
}
