% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chllogis}
\alias{chllogis}
\title{Log-logistic (LL) cumulative hazard function.}
\usage{
chllogis(t, mu, sigma)
}
\arguments{
\item{t}{: positive argument}

\item{mu}{: mean parameter in the log scale}

\item{sigma}{: scale parameter in the log scale}
}
\value{
the value of the LL cumulative hazard function
}
\description{
Log-logistic (LL) cumulative hazard function.
}
