% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{chlnorm}
\alias{chlnorm}
\title{Lognormal (LN) cumulative hazard function.}
\usage{
chlnorm(t, mu, sigma)
}
\arguments{
\item{t}{: positive argument}

\item{mu}{: mean parameter in the log scale}

\item{sigma}{: scale parameter in the log scale}
}
\value{
the value of the LN cumulative hazard function
}
\description{
Lognormal (LN) cumulative hazard function.
}
