% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{criterion_simu_wrapper}
\alias{criterion_simu_wrapper}
\title{Compare different selection criteria using simulated data}
\usage{
criterion_simu_wrapper(ind_wrapper, design, exhaust_ls, pen_vect,
  scaled_ground_haz)
}
\arguments{
\item{ind_wrapper}{index of the simulation design. Between 1 and \code{nrow(design)}.}

\item{design}{simulation design matrix}

\item{exhaust_ls}{list of exhaustive statistics computed from simulated data}

\item{pen_vect}{the vector of penalties}

\item{scaled_ground_haz}{the true hazard used to simulate data}
}
\value{
The estimated models from each criterion. The results are given in
the form of data frames. \code{nlevel} gives the number of selected areas,
\code{mse} gives the mean square error, \code{sel} gives the estimated regions,
and \©ode{haz} gives the estimated hazard.
}
\description{
Compare different selection criteria using simulated data
}
\seealso{
Other simu_wrapper: \code{\link{ac_simu_wrapper}},
  \code{\link{cv_simu_wrapper}}, \code{\link{design_simu}},
  \code{\link{ridge_simu_wrapper}}
}
