% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.R
\name{design_simu}
\alias{design_simu}
\title{Generate simulatied data}
\usage{
design_simu(sample_size_vect, n_rep, censoring_lim = c(75, 100))
}
\arguments{
\item{sample_size_vect}{vector of sample sizes}

\item{n_rep}{number of times the data are simulated for each sample size}

\item{censoring_lim}{interval of the censoring. Only cohort-indenpendant censoring is
available.}
}
\value{
sample_df simulated events

exhaust_ls list of exhaustive statistics computed for each simulation
}
\description{
Generate simulatied data
}
\seealso{
Other simu_wrapper: \code{\link{criterion_simu_wrapper}},
  \code{\link{cv_simu_wrapper}},
  \code{\link{ridge_simu_wrapper}}
}
