% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{dggamma}
\alias{dggamma}
\title{Generalised Gamma (GG) probability density function.
https://rpubs.com/FJRubio/GG}
\usage{
dggamma(t, sigma, nu, gamma, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the GG probability density function
}
\description{
Generalised Gamma (GG) probability density function.
https://rpubs.com/FJRubio/GG
}
