% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{EWMLE}
\alias{EWMLE}
\title{EWMLE function: Hazard Regression Models with Exponentiated Weibull (EW) baseline hazard}
\usage{
EWMLE(
  init,
  times,
  status,
  hstr = "EW",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (log(scale), log(shape1), log(shape2), alpha, beta)}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
EW without covariates ("EW"),
AFT model with EW baseline hazard ("EWAFT"),
PH model with EW baseline hazard ("EWPH"),
AH model with EW baseline hazard ("EWAH"),
GH model with EW baseline hazard ("EWGH")}

\item{des}{: design matrix for hazard-level effects}

\item{des_t}{: design matrix for time-level effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
EWMLE function: Hazard Regression Models with Exponentiated Weibull (EW) baseline hazard
}
