% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{GEHMLE}
\alias{GEHMLE}
\title{Relative (Net) Survival models
Log likelihood and MLE for the GH excess hazards model.
Baseline hazards: Lognormal, Log-logistic, Gamma, PGW, EW, GG}
\usage{
GEHMLE(
  init,
  times,
  status,
  hp,
  hstr = NULL,
  dist = NULL,
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{initial point for optimisation step}

\item{times}{times to event}

\item{status}{vital status indicators}

\item{hp}{population hazard}

\item{hstr}{hazard structure: "baseline", "AFT", "PH", "AH", or "GH"}

\item{dist}{distribution for the baseline hazard: "PGW", "GenGamma", "EW", "Weibull", "Gamma", "LogNormal", "LogLogistic"}

\item{des}{design matrix for hazard-level effects}

\item{des_t}{design matrix for time-level effects}

\item{method}{"nlminb" or optimisation method to be used in optim}

\item{maxit}{maximum number of iterations}
}
\value{
It returns the output from optim or nlminb and the negative log likelihood function
}
\description{
Relative (Net) Survival models
Log likelihood and MLE for the GH excess hazards model.
Baseline hazards: Lognormal, Log-logistic, Gamma, PGW, EW, GG
}
