% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{GGMLE}
\alias{GGMLE}
\title{GGMLE function: Hazard Regression Models with Generalised Gamma (GG) baseline hazard}
\usage{
GGMLE(
  init,
  times,
  status,
  hstr = "GG",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (log(scale), log(shape1), log(shape2), alpha, beta)}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
GG without covariates ("GG"),
AFT model with GG baseline hazard ("GGAFT"),
PH model with GG baseline hazard ("GGPH"),
AH model with GG baseline hazard ("GGAH"),
GH model with GG baseline hazard ("GGGH")}

\item{des}{: design matrix for hazard-level effects}

\item{des_t}{: design matrix for time-level effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
GGMLE function: Hazard Regression Models with Generalised Gamma (GG) baseline hazard
}
