% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{GHMLE}
\alias{GHMLE}
\title{GHMLE function: Hazard Regression Models with a parametric baseline hazard}
\usage{
GHMLE(
  init,
  times,
  status,
  hstr = NULL,
  dist = NULL,
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (log(scale), log(shape1), log(shape2), alpha, beta) for scale-shape1-shape2 models or
(mu, log(scale), alpha, beta) for log-location scale models.}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
 No covariates ("baseline"),
AFT model with PGW baseline hazard ("AFT"),
PH model with PGW baseline hazard ("PH"),
AH model with PGW baseline hazard ("AH"),
GH model with PGW baseline hazard ("GH")
*GH is not available with Weibull dist}

\item{dist}{: distribution for the baseline hazard:
Power Generalised Weibull ("PGW")
Generalised Gamma ("GenGamma"))
Exponentiated Weibull ("EW")
Weibull ("Weibull")
Gamma ("Gamma")
LogNormal ("LogNormal")
LogLogistic ("LogLogistic")}

\item{des}{: design matrix for hazard-level effects}

\item{des_t}{: design matrix for time-level effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
GHMLE function: Hazard Regression Models with a parametric baseline hazard
}
