% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{GMLE}
\alias{GMLE}
\title{GMLE function: Hazard Regression Models with Gamma (G) baseline hazard}
\usage{
GMLE(
  init,
  times,
  status,
  hstr = "G",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (log(scale), log(shape), alpha, beta)}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
G without covariates ("G"),
AFT model with G baseline hazard ("GAFT"),
PH model with G baseline hazard ("GPH"),
AH model with G baseline hazard ("GAH"),
GH model with G baseline hazard ("GGH")}

\item{des}{: design matrix for hazard-level effects}

\item{des_t}{: design matrix for time-level effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
GMLE function: Hazard Regression Models with Gamma (G) baseline hazard
}
