% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{package}
\name{hazreg}
\alias{hazreg}
\alias{hazreg-package}
\title{foo: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Utility functions}{
NA
}

\section{The Age-cohort model}{
NA
}

\section{The Age-cohort interaction model}{
NA
}

\section{The interaction model}{
NA
}

