% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{hessian_interaction}
\alias{hessian_interaction}
\title{Second order derivate of the negative lok-likelihood in the interaction model}
\usage{
hessian_interaction(par, O, R, pen, weights_age = NULL,
  weights_cohort = NULL, use_band = FALSE)
}
\arguments{
\item{par}{parameters of the interaction model in the form \code{c(mu, alpha and beta)}}

\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}
}
\value{
The matrix of second order derivatives of the negative log-likelihood as
described in TODO NAME OF REFERENCE SHEET
}
\description{
Second order derivate of the negative lok-likelihood in the interaction model
}
\examples{
J <- 10
K <- 15
set.seed(0)
O <- matrix(rpois(K * J, 2), K, J)
R <- matrix(rpois(K * J, 10), K, J)
par <- rnorm(J * K)
hessian_interaction(par, O, R)
}
\seealso{
Other ac_likelihood: \code{\link{hessian_ac}},
  \code{\link{loglik_ac}}, \code{\link{score_ac}},
  \code{\link{solver_ac}}
}
