% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{hew}
\alias{hew}
\title{Power Exponentiated Weibull (EW) hazard function.
https://rpubs.com/FJRubio/EWD}
\usage{
hew(t, sigma, nu, gamma, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the EW hazard function
}
\description{
Power Exponentiated Weibull (EW) hazard function.
https://rpubs.com/FJRubio/EWD
}
