% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{hgamma}
\alias{hgamma}
\title{Gamma (G) hazard function.}
\usage{
hgamma(t, shape, scale, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{shape}{: shape parameter}

\item{scale}{: scale parameter}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the Gamma hazard function
}
\description{
Gamma (G) hazard function.
}
