% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{hlnorm}
\alias{hlnorm}
\title{Lognormal (LN) hazard function.}
\usage{
hlnorm(t, mu, sigma, log = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{mu}{: mean parameter in the log scale}

\item{sigma}{: scale parameter in the log scale}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the LN hazard function
}
\description{
Lognormal (LN) hazard function.
}
