% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{LNMLE}
\alias{LNMLE}
\title{LNMLE function: Hazard Regression Models with Lognormal (LN) baseline hazard}
\usage{
LNMLE(
  init,
  times,
  status,
  hstr = "LN",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (meanlog, log(scale), alpha, beta)}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
LN without covariates ("LN"),
AFT model with LN baseline hazard ("LNAFT"),
PH model with LN baseline hazard ("LNPH"),
AH model with LN baseline hazard ("LNAH"),
GH model with LN baseline hazard ("LNGH")}

\item{des}{: design matrix for hazard-level effects}

\item{des_t}{: design matrix for time-level effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
LNMLE function: Hazard Regression Models with Lognormal (LN) baseline hazard
}
