% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{loglik_aci_sel}
\alias{loglik_aci_sel}
\alias{score_aci_sel}
\alias{hessian_aci_sel}
\title{Negative lok-likelihood in the age-cohort-interaction model with constaints}
\usage{
loglik_aci_sel(par, O, R, sel_array)

score_aci_sel(par, O, R, sel_array)

hessian_aci_sel(par, O, R, sel_array)
}
\arguments{
\item{par}{parameter of the age-cohort-interaction model in the
form \code{c(mu, alpha and beta)}}

\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}

\item{sel_array}{an array defining the selection found by the adaptive ridge
procedure. (see \code{aridge_solver_aci})}
}
\value{
The vector of derivatives of the negative log-likelihood as
described in TODO NAME OF REFERENCE SHEET
}
\description{
Negative lok-likelihood in the age-cohort-interaction model with constaints
}
\examples{
\dontrun{
par <- rnorm(ncol(O) * nrow(O))
loglik_aci_sel(par, O, R) # length 1 vector
score_aci_sel(par, O, R) # vector of length (ncol(O) + nrow(O))
hessian_aci_sel(par, O, R) # square matrix of dimension (ncol(O) + nrow(O))
}
}
