% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{loglik_interaction}
\alias{loglik_interaction}
\title{Negative lok-likelihood in the interaction model}
\usage{
loglik_interaction(par, O, R, pen, weights_age = NULL,
  weights_cohort = NULL)
}
\arguments{
\item{par}{parameters of the interaction model in the form \code{c(mu, alpha and beta)}}

\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}
}
\value{
The negative log-likelihood as described in TODO NAME OF REFERENCE SHEET
}
\description{
Negative lok-likelihood in the interaction model
}
\examples{
\dontrun{
J <- 10
K <- 15
set.seed(0)
O <- matrix(rpois(K * J, 2), K, J)
R <- matrix(rpois(K * J, 10), K, J)
par <- rnorm(J * K)
loglik_interaction(par, O, R)
}
}
\concept{interaction_likelihood}
