% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{map_surv}
\alias{map_surv}
\title{Generate age and cohort observations}
\usage{
map_surv(map, sample_size, dob_dist, dob_dist_par = NULL)
}
\arguments{
\item{map}{A matrix representing the hazard, as provided by \code{\link{map_maker}}.}

\item{sample_size}{Number of observations}

\item{dob_dist}{Marginal distribution of the cohort. Defaults to the uniform distribution
with bounds given by \code{dob_dist_par}.}

\item{dob_dist_par}{Parameter of the distribution of the cohort}
}
\value{
\code{surv_data}, a data frame of observations with age and cohort.
}
\description{
Generate age and cohort observations
}
\examples{
library(tidyverse)
mu <- log(1e-2)
age_seq <- seq(0, 100, length.out = 100)
cohort_seq <- seq(1900, 1900 + age_seq[101], length.out = age_length + 1)
age_coef <- seq(0, 2, length.out = age_length - 1)
cohort_coef <- seq(0, 0.8, length.out = age_length - 1)
islet_ls = list(list(mean = c(1945, 45), sigma = 100 * diag(2), coef = 40))
hazard <- map_maker(mu, age_coef, cohort_coef, age_seq, cohort_seq, islet_ls = islet_ls)
(map_surv(hazard, 10))
}
