% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{par2haz_ac}
\alias{par2haz_ac}
\title{Create hazard from parameters of the Age-Cohort Model}
\usage{
par2haz_ac(par, J, K)
}
\arguments{
\item{par}{parameters of the age-cohort model in the form \code{c(mu, alpha and beta)}}

\item{J}{number of age intervals}

\item{K}{number of cohort intervals}
}
\value{
The hazard rate in a matrix of size K * J:
\deqn{\lambda_{a, c} = \exp (\mu + \alpha_{a} + \beta_{c})}
}
\description{
Create hazard from parameters of the Age-Cohort Model
}
\examples{
J <- 10
K <- 15
par <- rnorm(K + J - 1)
persp(par2haz_ac(par, J, K))
}
