% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{par2haz_aci}
\alias{par2haz_aci}
\alias{par2grid_aci}
\alias{par_sel2par_aci}
\title{Transform parameter to hazard in the age-cohort-interaction model}
\usage{
par2haz_aci(par, J, K)

par2grid_aci(par, cuts_age, cuts_cohort)

par_sel2par_aci(par_sel, sel)
}
\arguments{
\item{par}{parameter of the age-cohort-interaction model in the
form \code{c(mu, alpha and beta)}}

\item{J}{length of \code{cuts_age}}

\item{K}{length of \code{cuts_cohort}}
}
\value{
a matrix of dimension \code{K * J}
}
\description{
Transform parameter to hazard in the age-cohort-interaction model
}
\examples{
\dontrun{
J <-  10
K <-  15
set.seed(0)
persp(par2haz_aci(rnorm(J * K), K, J))
}
}
