% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{PGWMLE}
\alias{PGWMLE}
\title{PGWMLE function: Hazard Regression Models with Power Generalised Weibull (PGW) baseline hazard}
\usage{
PGWMLE(
  init,
  times,
  status,
  hstr = "PGW",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (log(scale), log(shape1), log(shape2), alpha, beta)}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
PGW without covariates ("PGW"),
AFT model with PGW baseline hazard ("PGWAFT"),
PH model with PGW baseline hazard ("PGWPH"),
AH model with PGW baseline hazard ("PGWAH"),
GH model with PGW baseline hazard ("PGWGH")}

\item{des}{: design matrix for hazard-level effects}

\item{des_t}{: design matrix for time-level effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
PGWMLE function: Hazard Regression Models with Power Generalised Weibull (PGW) baseline hazard
}
