% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{qggamma}
\alias{qggamma}
\title{Generalised Gamma (GG) quantile function.
https://rpubs.com/FJRubio/GG}
\usage{
qggamma(p, sigma, nu, gamma)
}
\arguments{
\item{p}{:     probability. A value in (0,1)}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}
}
\value{
the value of the GG quantile function
}
\description{
Generalised Gamma (GG) quantile function.
https://rpubs.com/FJRubio/GG
}
