% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{qllogis}
\alias{qllogis}
\title{Log-logistic (LL) quantile function.}
\usage{
qllogis(p, mu, sigma)
}
\arguments{
\item{p}{:     probability. A value in (0,1)}

\item{mu}{: mean parameter in the log scale}

\item{sigma}{: scale parameter in the log scale}

\item{log:}{log scale (TRUE or FALSE)}
}
\value{
the value of the LL quantile function
}
\description{
Log-logistic (LL) quantile function.
}
