% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{qpgw}
\alias{qpgw}
\title{Power Generalised Weibull (PGW) quantile function.
http://rpubs.com/FJRubio/PGW}
\usage{
qpgw(p, sigma, nu, gamma)
}
\arguments{
\item{p}{:     probability. A value in (0,1)}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}
}
\value{
the value of the PGW quantile function
}
\description{
Power Generalised Weibull (PGW) quantile function.
http://rpubs.com/FJRubio/PGW
}
