% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{raster}
\alias{raster}
\title{Color plot a matrix built on top of \code{ggplot2}'s \code{geom_raster}}
\usage{
raster(mat, type = c("factor", "numeric"))
}
\arguments{
\item{mat}{matrix to be plotted}

\item{type}{whether \code{mat}'s values represent a quantitative (\code{numeric})
or qualitative (\code{factor}) variable}
}
\value{
The ggplot element is returned as \code{\link{invisible}}. It is also displayed
in the graphical window
}
\description{
Color plot a matrix built on top of \code{ggplot2}'s \code{geom_raster}
}
