% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{ridge_solver_aci}
\alias{ridge_solver_aci}
\alias{aridge_solver_aci}
\title{Computes the ridge regularized estimates of the age-cohort-interaction
 model}
\usage{
ridge_solver_aci(O, R, pen, weights_age = NULL, weights_cohort = NULL,
  old_par = NULL, use_band = FALSE, maxiter = 1000)

aridge_solver_aci(O, R, pen_vect, sample_size, use_band = FALSE,
  maxiter = 1000 * length(pen_vect))
}
\arguments{
\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}

\item{pen}{The penalty constant.}

\item{weights_age}{weights of the age differences in the log-hazard. See ADD REFERENCE SHEET}

\item{weights_cohort}{weights of the cohort differences in the log-hazard. See ADD REFERENCE SHEET}

\item{old_par}{initial point of the iteration. The default choice is \code{0}
is recommended}

\item{use_band}{Whether to use faster inversion of the hessian function. \code{TRUE}}

\item{maxiter}{The maximal number of iteractions of the Newton-Raphson procedure.}

\item{verbose}{Whether to display the progress of the Newton-Raphson procedure.}
}
\value{
The vector estimate of the ridge regularized age-cohort-interaction model.
}
\description{
Computes the ridge regularized estimates of the age-cohort-interaction
 model
}
\examples{
\dontrun{
J <- 10
K <- 15
set.seed(0)
O <- matrix(rpois(K * J, 2), K, J)
R <- matrix(rpois(K * J, 10), K, J)
pen <-  50
ridge <- ridge_solver_aci(O, R, pen)
ridge$par
ridge$iter
}
}
\seealso{
Other ridge: \code{\link{ridge_solver_interaction}}

Other adaptive_ridge: \code{\link{ridge_solver_interaction}}
}
