% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aci.R
\name{ridge_solver_aci_sel}
\alias{ridge_solver_aci_sel}
\title{Computes the estimate of the constrained aci model}
\usage{
ridge_solver_aci_sel(O, R, sel_array, maxiter = 1000)
}
\arguments{
\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}

\item{sel_array}{an array defining the selection found by the adaptive ridge
procedure. (see \code{aridge_solver_aci})}
}
\value{
The vector estimate of the constrained age-cohort-interaction model.
The \code{sel_array} is an array of dimension \code{K  J  L}
}
\description{
Computes the estimate of the constrained aci model
}
