% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pch.R
\name{rpch}
\alias{rpch}
\title{Generate sample from the piecewise constant hazard model}
\usage{
rpch(n, cuts, alpha)
}
\arguments{
\item{n}{sample size}

\item{cuts}{cuts defining the intervals in the piecewise constant hazard model}

\item{alpha}{values of the hazard (over each interval) in the piecewise constant hazard model}
}
\value{
a matrix of dimension \code{K * J}
}
\description{
Generate sample from the piecewise constant hazard model
}
\examples{
\dontrun{
data <- rpch(1000, 1, c(0.2, 0.9))
plot(ecdf(data))
}
}
