% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{rpgw}
\alias{rpgw}
\title{Power Generalised Weibull (PGW) random number generation.
http://rpubs.com/FJRubio/PGW}
\usage{
rpgw(n, sigma, nu, gamma)
}
\arguments{
\item{n}{:     number of observations}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}
}
\value{
generates random deviates
}
\description{
Power Generalised Weibull (PGW) random number generation.
http://rpubs.com/FJRubio/PGW
}
