% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rsurv}
\alias{rsurv}
\title{Generate data from piecewise constant hazard}
\usage{
rsurv(n, cuts, alpha)
}
\arguments{
\item{n}{number of observations. A positive integer}

\item{cuts}{cuts of the piecewise constant hazard}

\item{alpha}{values of the piecewise constant hazard}
}
\value{
A sequence of observations of length \code{n} generated from the
distribution given by \code{cuts} and \code{alpha}
}
\description{
Generate data from piecewise constant hazard
}
\examples{
\dontrun{
map_maker(100, c(20, 40, 60), c(0.01, 0.03, 0.05, 0.08))
}
}
