% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interaction.R
\name{score_interaction}
\alias{score_interaction}
\title{First order derivate of the negative lok-likelihood in the interaction model}
\usage{
score_interaction(par, O, R, pen, weights_age = NULL,
  weights_cohort = NULL)
}
\arguments{
\item{par}{parameters of the interaction model in the form \code{as.vector(eta)}}

\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}
}
\value{
The vector of derivatives of the negative log-likelihood as
described in TODO NAME OF REFERENCE SHEET
}
\description{
First order derivate of the negative lok-likelihood in the interaction model
}
\examples{
\dontrun{
J <- 10
K <- 15
set.seed(0)
O <- matrix(rpois(K * J, 2), K, J)
R <- matrix(rpois(K * J, 10), K, J)
par <- rnorm(J * K)
score_interaction(par, O, R)
}
}
\concept{likelihood_likelihood}
