% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazreg.R
\name{seq_geom}
\alias{seq_geom}
\title{Generate geometric sequences}
\usage{
seq_geom(a, b, length.out)
}
\arguments{
\item{length.out}{The length of the sequence. It must be a strictly positive integer}

\item{x}{A strictly positive number}

\item{y}{A strictly positive number}
}
\value{
A geometric sequence
}
\description{
Generate geometric sequences
}
\examples{
seq_geom(1, 10, 10)
seq_geom(1, 10, 10)
}
