% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{sggamma}
\alias{sggamma}
\title{Generalised Gamma (GG) survival function.
https://rpubs.com/FJRubio/GG}
\usage{
sggamma(t, sigma, nu, gamma, log.p = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}

\item{log.p:}{log scale (TRUE or FALSE)}
}
\value{
the value of the GG survival function
}
\description{
Generalised Gamma (GG) survival function.
https://rpubs.com/FJRubio/GG
}
