% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{simGH}
\alias{simGH}
\title{simGH function: Function to simulate times to event from a model with a GH structure
for different parametric baseline hazards.
Distributions: LN, LL, GG, G (Gamma), W (Weibull), PGW, EW.
See: https://github.com/FJRubio67/HazReg}
\usage{
simGH(
  seed,
  n,
  des = NULL,
  des_h = NULL,
  des_t = NULL,
  theta,
  beta_h = NULL,
  beta_t = NULL,
  beta = NULL,
  hstr,
  baseline
)
}
\arguments{
\item{seed}{: seed for simulation}

\item{n}{: sample size (number of individuals)}

\item{des}{: Design matrix for AFT, PH, and AH models}

\item{des_h}{: Design matrix for GH model (hazard scale)}

\item{des_t}{: Design matrix for GH model (time scale)}

\item{theta}{:  parameters of the baseline hazard}

\item{beta_h}{: regression parameters multiplying the hazard for GH model}

\item{beta_t}{: regression parameters multiplying the time scale for GH model}

\item{beta}{: regression parameters for AFT, PH, and AH models}

\item{hstr}{: hazard structure (AH, AFT, PH, GH)}

\item{baseline}{: baseline hazard distribution}
}
\value{
a vector containing the simulated times to event
}
\description{
simGH function: Function to simulate times to event from a model with a GH structure
for different parametric baseline hazards.
Distributions: LN, LL, GG, G (Gamma), W (Weibull), PGW, EW.
See: https://github.com/FJRubio67/HazReg
}
