% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{spgw}
\alias{spgw}
\title{Power Generalised Weibull (PGW) survival function.
http://rpubs.com/FJRubio/PGW}
\usage{
spgw(t, sigma, nu, gamma, log.p = FALSE)
}
\arguments{
\item{t}{: positive argument}

\item{sigma}{: scale parameter}

\item{nu}{: shape parameter}

\item{gamma}{: shape parameter}

\item{log.p:}{log scale (TRUE or FALSE)}
}
\value{
the value of the PGW survival function
}
\description{
Power Generalised Weibull (PGW) survival function.
http://rpubs.com/FJRubio/PGW
}
