% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routines.R
\name{WMLE}
\alias{WMLE}
\title{WMLE function: Hazard Regression Models with Weibull (W) baseline hazard}
\usage{
WMLE(
  init,
  times,
  status,
  hstr = "W",
  des = NULL,
  des_t = NULL,
  method = "Nelder-Mead",
  maxit = 100
)
}
\arguments{
\item{init}{: initial point for optimisation step
under the parameterisation (log(scale), log(shape), alpha, beta)}

\item{times}{: times to event}

\item{status}{: vital status indicators (TRUE or 1 = observed, FALSE  or 0 = censored)}

\item{hstr}{: hazard structure:
W without covariates ("W"),
AFT model with W baseline hazard ("WAFT"),
PH model with W baseline hazard ("WPH"),
AH model with W baseline hazard ("WAH")}

\item{des}{: design matrix for hazard-level effects}

\item{des_t}{: design matrix for time-level effects (it is recommended not to use splines here)}

\item{method}{: "nlminb" or optimisation method to be used in optim (see ?optim)}

\item{maxit}{: maximum number of iterations in optim or nlminb}
}
\value{
It returns the output from optim or nlminb for the selected model and the negative log likelihood function
}
\description{
WMLE function: Hazard Regression Models with Weibull (W) baseline hazard
}
