% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazreg.R
\name{band_prod}
\alias{band_prod}
\title{Solve linear system when the matrix is block-wise with a band diagonal block}
\usage{
band_prod(hessian_mat, score_vect, lim)
}
\arguments{
\item{hessian_mat}{Square matrix}

\item{score_vect}{Vector of same length as \code{hessian_mat}}

\item{lim}{Positive integer}
}
\value{
The vector \code{hessian_mat} ^ {-1} \code{score_vect} is returned.
If \code{hessian_mat[lim:end, lim:end]} is not a band matrix or if \code{hessian_mat} is
symmetric, an error is returned.
If these two conditions are met, the returned value is the same as
\code{solve(hessian_mat, score_vect)}. There might be nummerical impercision errors.
}
\description{
Solve linear system when the matrix is block-wise with a band diagonal block
}
\examples{
set.seed(0)
A <- matrix(rnorm(36), 6, 6)
B <- matrix(rnorm(24), 6, 4)
C <- diag(rnorm(4)) + rbind(cbind(0, diag(rnorm(3))), 0)
hessian_mat <- rbind(cbind(A + t(A), B),
                     cbind(t(B), C + t(C)))
score_vect <- rnorm(10)
lim <- 6
max(band_prod(hessian_mat, score_vect, lim) - solve(hessian_mat, score_vect))
}
