% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\docType{package}
\name{hazreg}
\alias{hazreg}
\alias{hazreg-package}
\title{hazreg: A package for computing regularized hazard in survival analysis models}
\description{
The foo package provides functions for the estimation of hazard
of different models in time-to-event data.
}
\section{Utility functions}{
NA
}

\section{The Age-cohort model}{
NA
}

\section{The Age-cohort interaction model}{
NA
}

\section{The interaction model}{
NA
}

