% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{loglik_ac}
\alias{loglik_ac}
\title{Negative lok-likelihood in the age-cohort model}
\usage{
loglik_ac(par, O, R)
}
\arguments{
\item{par}{parameters of the age-cohort model in the form \code{c(mu, alpha and beta)}}

\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}
}
\value{
The negative log-likelihood as described in TODO NAME OF REFERENCE SHEET
}
\description{
Negative lok-likelihood in the age-cohort model
}
\examples{
\dontrun{
J <- 10
K <- 15
set.seed(0)
O <- matrix(rpois(K * J, 2), K, J)
R <- matrix(rpois(K * J, 10), K, J)
par <- rnorm(J + K - 1)
loglik_ac(par, O, R)
}
}
\seealso{
Other ac_likelihood: \code{\link{hessian_ac}},
  \code{\link{hessian_interaction}},
  \code{\link{score_ac}}, \code{\link{solver_ac}}
}
\concept{ac_likelihood}
