% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazreg.R
\name{loglik_aci}
\alias{loglik_aci}
\title{Negative lok-likelihood in the age-cohort-interaction model}
\usage{
loglik_aci(par, O, R, pen, weights_age = NULL, weights_cohort = NULL)
}
\arguments{
\item{par}{parameters of the age-cohort model in the form \code{c(mu, alpha and beta)}}

\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}
}
\value{
The vector of derivatives of the negative log-likelihood as
described in TODO NAME OF REFERENCE SHEET
}
\description{
Negative lok-likelihood in the age-cohort-interaction model
}
\examples{
par <- rnorm(ncol(O) + nrow(O))
loglik_ac(par, O, R) # length 1 vector
score_ac(par, O, R) # vector of length (ncol(O) + nrow(O))
hessian_ac(par, O, R) # square matrix of dimension (ncol(O) + nrow(O))
}
