% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{score_ac}
\alias{score_ac}
\title{First order derivate of the negative lok-likelihood in the age-cohort model}
\usage{
score_ac(par, O, R)
}
\arguments{
\item{par}{parameters of the age-cohort model in the form \code{c(mu, alpha and beta)}}

\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}
}
\value{
The vector of derivatives of the negative log-likelihood as
described in TODO NAME OF REFERENCE SHEET
}
\description{
First order derivate of the negative lok-likelihood in the age-cohort model
}
\examples{
par <- rnorm(ncol(O) + nrow(O))
score_ac(par, O, R)
}
