% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ac.R
\name{solver_ac}
\alias{solver_ac}
\title{Solver of the Age-cohort model}
\usage{
solver_ac(O, R, maxiter = 1000, verbose = FALSE)
}
\arguments{
\item{O}{Observed events as returned by \code{\link{exhaustive_stat_2d}}}

\item{R}{Time at risk as returned by \code{\link{exhaustive_stat_2d}}}
}
\description{
Solver of the Age-cohort model
}
\examples{
par <- rnorm(ncol(O) + nrow(O))
hessian_ac(par, O, R)
}
\seealso{
\code{\link{exhaustive_stat_2d}}, \code{\link{exhaustive_stat_1d}}

Other ac_likelihood: \code{\link{hessian_ac}},
  \code{\link{hessian_interaction}},
  \code{\link{loglik_ac}}, \code{\link{score_ac}}
}
